Rename formatCurrency to format
