+++ packages/blockchain-utils/source/index.ts
@@ -1,4 +1,4 @@
-export { formatCurrency } from "./format-currency.js";
+export { formatCurrency } from "./format.js";
 export { getBlockNotChainedErrorMessage, isBlockChained } from "./is-block-chained.js";
 export { isMajority } from "./is-majority.js";
 export { isMinority } from "./is-minority.js";
